+++ packages/polling-controller/src/index.ts
@@ -6,6 +6,7 @@ export {
 export {
   StaticIntervalPollingControllerOnly,
   StaticIntervalPollingController,
+  StaticIntervalPollingControllerNext,
 } from './StaticIntervalPollingController';
 
 export type { IPollingController } from './types';